+++ public/wp-content/plugins/disable-emails/includes/bootstrap.php
@@ -12,6 +12,7 @@ require DISABLE_EMAILS_PLUGIN_ROOT . 'includes/functions.php';
 // replace standard WordPress wp_mail() if nobody else has already done it
 if (!function_exists('wp_mail')) {
 
+	require DISABLE_EMAILS_PLUGIN_ROOT . 'includes/class.EmailAddress.php';
 	require DISABLE_EMAILS_PLUGIN_ROOT . 'includes/class.PHPMailerMock.php';
 
 	function wp_mail( $to, $subject, $message, $headers = '', $attachments = [] ) {
@@ -25,7 +26,7 @@ if (!function_exists('wp_mail')) {
 }
 
 /**
-* kick start the plugin
-*/
+ * kick start the plugin
+ */
 require DISABLE_EMAILS_PLUGIN_ROOT . 'includes/class.Plugin.php';
 Plugin::getInstance()->pluginStart();