+++ public/wp-content/plugins/disable-emails/includes/functions.php
@@ -2,7 +2,7 @@
 
 namespace webaware\disable_emails;
 
-use \Exception;
+use Exception;
 
 if (!defined('ABSPATH')) {
 	exit;
@@ -10,10 +10,15 @@ if (!defined('ABSPATH')) {
 
 const OPT_SETTINGS				= 'disable_emails';
 
+const INDICATOR_NONE			= 'none';
+const INDICATOR_TOOLBAR			= 'toolbar';
+const INDICATOR_NOTICE			= 'notice';
+const INDICATOR_NOTICE_AND_TB	= 'notice_toolbar';
+
 /**
-* get current plugin settings, use defaults if settings not yet saved
-* @return array
-*/
+ * get current plugin settings, use defaults if settings not yet saved
+ * @return array
+ */
 function get_plugin_settings() {
 	$defaults = [
 		'wp_mail'				=> 1,
@@ -24,25 +29,26 @@ function get_plugin_settings() {
 		'phpmailer_init'		=> 1,
 		'buddypress'			=> 1,
 		'events_manager'		=> 1,
+		'indicator'				=> INDICATOR_TOOLBAR,
 	];
 
-	return get_option(OPT_SETTINGS, $defaults);
+	return wp_parse_args(get_option(OPT_SETTINGS, []), $defaults);
 }
 
 /**
-* can current user activate/deactivate the must-use plugin?
-* @return bool
-*/
+ * can current user activate/deactivate the must-use plugin?
+ * @return bool
+ */
 function has_mu_plugin_permission() {
 	return current_user_can(is_multisite() ? 'manage_network_plugins' : 'activate_plugins');
 }
 
 /**
-* install, update, or remove the must-use plugin
-* @param string $action
-* @return bool
-* @throws Exception
-*/
+ * install, update, or remove the must-use plugin
+ * @param string $action
+ * @return bool
+ * @throws Exception
+ */
 function mu_plugin_manage($action) {
 	if (!has_mu_plugin_permission()) {
 		throw new Exception(__('No permission to manage Disable Emails must-use plugin.', 'disable-emails'));
@@ -50,8 +56,17 @@ function mu_plugin_manage($action) {
 
 	$has_mu_plugin = defined('DISABLE_EMAILS_MU_PLUGIN');
 
-	$source = wp_normalize_path(DISABLE_EMAILS_PLUGIN_ROOT . '/mu-plugin/disable-emails-mu.php');
-	$target = wp_normalize_path(WPMU_PLUGIN_DIR . '/disable-emails-mu.php');
+	$wpmu_plugin_dir = rtrim(wp_normalize_path(WPMU_PLUGIN_DIR), '/');
+	if (!is_dir($wpmu_plugin_dir)) {
+		// folder does not exist, create it now
+		wp_mkdir_p($wpmu_plugin_dir);
+		if (!is_dir($wpmu_plugin_dir)) {
+			throw new Exception(__('Unable to create folder for Disable Emails must-use plugin.', 'disable-emails'));
+		}
+	}
+
+	$source = wp_normalize_path(DISABLE_EMAILS_PLUGIN_ROOT . 'mu-plugin/disable-emails-mu.php');
+	$target = "$wpmu_plugin_dir/disable-emails-mu.php";
 
 	switch ($action) {
 
@@ -76,4 +91,27 @@ function mu_plugin_manage($action) {
 	}
 
 	return $has_mu_plugin;
+}
+
+/**
+ * get message for current active status
+ * @return string
+ */
+function get_status_message() {
+	require_once ABSPATH . 'wp-admin/includes/plugin.php';
+
+	if (defined('DISABLE_EMAILS_MU_PLUGIN') && is_multisite()) {
+		/* translators: shown when emails are disabled for all sites in all networks in a multisite, with the must-use plugin */
+		$msg = __('Emails are disabled for all sites.', 'disable-emails');
+	}
+	elseif (is_plugin_active_for_network(DISABLE_EMAILS_PLUGIN_NAME)) {
+		/* translators: shown when emails are disabled for all sites in a multisite network, by network-activating the plugin */
+		$msg = __('Emails are disabled on this network.', 'disable-emails');
+	}
+	else {
+		/* translators: shown when emails are disabled for the current site */
+		$msg = __('Emails are disabled.', 'disable-emails');
+	}
+
+	return $msg;
 }