+++ public/wp-content/plugins/radio-buttons-for-taxonomies/js/src/radio-term-selector.js
@@ -1,18 +1,30 @@
-/*
+/**
  * Modified version of https://github.com/WordPress/gutenberg/blob/master/packages/editor/src/components/post-taxonomies/hierarchical-term-selector.js
  */
 
 /**
  * External dependencies
  */
-import { get, unescape as unescapeString, without, find, some, invoke } from 'lodash';
+import {
+	get,
+	unescape as unescapeString,
+	without,
+	find,
+	some,
+	invoke,
+} from 'lodash';
 
 /**
  * WordPress dependencies
  */
 import { __, _x, _n, sprintf } from '@wordpress/i18n';
 import { Component } from '@wordpress/element';
-import { TreeSelect, withSpokenMessages, withFilters, Button } from '@wordpress/components';
+import {
+	RadioControl,
+	TreeSelect,
+	withSpokenMessages,	
+	Button,
+} from '@wordpress/components';
 import { withSelect, withDispatch } from '@wordpress/data';
 import { withInstanceId, compose } from '@wordpress/compose';
 import apiFetch from '@wordpress/api-fetch';
@@ -59,14 +71,19 @@ class RadioTermSelector extends Component {
 		};
 	}
 
-	onChange( event ) { // @helgatheviking
+	onChange( termId ) { // @helgatheviking
 		const { onUpdateTerms, taxonomy } = this.props;
-		const termId = parseInt( event.target.value, 10 );
 		onUpdateTerms( [ termId ], taxonomy.rest_base );
 	}
 
+	onClear() { // @helgatheviking props @tomjn
+		const { onUpdateTerms, taxonomy } = this.props;
+		onUpdateTerms( [], taxonomy.rest_base );
+	}
+
 	onChangeFormName( event ) {
-		const newValue = event.target.value.trim() === '' ? '' : event.target.value;
+		const newValue =
+			event.target.value.trim() === '' ? '' : event.target.value;
 		this.setState( { formName: newValue } );
 	}
 
@@ -82,8 +99,11 @@ class RadioTermSelector extends Component {
 
 	findTerm( terms, parent, name ) {
 		return find( terms, ( term ) => {
-			return ( ( ! term.parent && ! parent ) || parseInt( term.parent ) === parseInt( parent ) ) &&
-				term.name.toLowerCase() === name.toLowerCase();
+			return (
+				( ( ! term.parent && ! parent ) ||
+					parseInt( term.parent ) === parseInt( parent ) ) &&
+				term.name.toLowerCase() === name.toLowerCase()
+			);
 		} );
 	}
 
@@ -96,11 +116,18 @@ class RadioTermSelector extends Component {
 		}
 
 		// check if the term we are adding already exists
-		const existingTerm = this.findTerm( availableTerms, formParent, formName );
+		const existingTerm = this.findTerm(
+			availableTerms,
+			formParent,
+			formName
+		);
 		if ( existingTerm ) {
 			// if the term we are adding exists but is not selected select it
 			if ( ! some( terms, ( term ) => term === existingTerm.id ) ) {
-				onUpdateTerms( [ existingTerm.id ], taxonomy.rest_base ); // @helgatheviking
+				onUpdateTerms(
+					[ existingTerm.id ], // @helgatheviking
+					taxonomy.rest_base
+				);
 			}
 			this.setState( {
 				formName: '',
@@ -121,29 +148,34 @@ class RadioTermSelector extends Component {
 			},
 		} );
 		// Tries to create a term or fetch it if it already exists
-		const findOrCreatePromise = this.addRequest
-			.catch( ( error ) => {
-				const errorCode = error.code;
-				if ( errorCode === 'term_exists' ) {
-					// search the new category created since last fetch
-					this.addRequest = apiFetch( {
-						path: addQueryArgs(
-							`/wp/v2/${ taxonomy.rest_base }`,
-							{ ...DEFAULT_QUERY, parent: formParent || 0, search: formName }
-						),
-					} );
-					return this.addRequest
-						.then( ( searchResult ) => {
-							return this.findTerm( searchResult, formParent, formName );
-						} );
-				}
-				return Promise.reject( error );
-			} );
-		findOrCreatePromise
-			.then( ( term ) => {
-				const hasTerm = !! find( this.state.availableTerms, ( availableTerm ) => availableTerm.id === term.id );
-				const newAvailableTerms = hasTerm ? this.state.availableTerms : [ term, ...this.state.availableTerms ];
+		const findOrCreatePromise = this.addRequest.catch( ( error ) => {
+			const errorCode = error.code;
+			if ( errorCode === 'term_exists' ) {
+				// search the new category created since last fetch
+				this.addRequest = apiFetch( {
+					path: addQueryArgs( `/wp/v2/${ taxonomy.rest_base }`, {
+						...DEFAULT_QUERY,
+						parent: formParent || 0,
+						search: formName,
+					} ),
+				} );
+				return this.addRequest.then( ( searchResult ) => {
+					return this.findTerm( searchResult, formParent, formName );
+				} );
+			}
+			return Promise.reject( error );
+		} );
+		findOrCreatePromise.then(
+			( term ) => {
+				const hasTerm = !! find(
+					this.state.availableTerms,
+					( availableTerm ) => availableTerm.id === term.id
+				);
+				const newAvailableTerms = hasTerm
+					? this.state.availableTerms
+					: [ term, ...this.state.availableTerms ];
 				const termAddedMessage = sprintf(
+					/* translators: %s: taxonomy name */
 					_x( '%s added', 'term' ),
 					get(
 						this.props.taxonomy,
@@ -158,10 +190,13 @@ class RadioTermSelector extends Component {
 					formName: '',
 					formParent: '',
 					availableTerms: newAvailableTerms,
-					availableTermsTree: this.sortBySelected( buildTermsTree( newAvailableTerms ) ),
+					availableTermsTree: this.sortBySelected(
+						buildTermsTree( newAvailableTerms )
+					),
 				} );
 				onUpdateTerms( [ term.id ], taxonomy.rest_base ); // @helgatheviking
-			}, ( xhr ) => {
+			},
+			( xhr ) => {
 				if ( xhr.statusText === 'abort' ) {
 					return;
 				}
@@ -169,7 +204,8 @@ class RadioTermSelector extends Component {
 				this.setState( {
 					adding: false,
 				} );
-			} );
+			}
+		);
 	}
 
 	componentDidMount() {
@@ -193,11 +229,18 @@ class RadioTermSelector extends Component {
 			return;
 		}
 		this.fetchRequest = apiFetch( {
-			path: addQueryArgs( `/wp/v2/${ taxonomy.rest_base }`, DEFAULT_QUERY ),
+			path: addQueryArgs(
+				`/wp/v2/${ taxonomy.rest_base }`,
+				DEFAULT_QUERY
+			),
 		} );
 		this.fetchRequest.then(
-			( terms ) => { // resolve
-				const availableTermsTree = this.sortBySelected( buildTermsTree( terms ) );
+			( terms ) => {
+				// resolve
+
+				const availableTermsTree = this.sortBySelected(
+					buildTermsTree( terms )
+				);
 
 				this.fetchRequest = null;
 				this.setState( {
@@ -206,7 +249,8 @@ class RadioTermSelector extends Component {
 					availableTerms: terms,
 				} );
 			},
-			( xhr ) => { // reject
+			( xhr ) => {
+				// reject
 				if ( xhr.statusText === 'abort' ) {
 					return;
 				}
@@ -227,7 +271,10 @@ class RadioTermSelector extends Component {
 			if ( undefined === termTree.children ) {
 				return false;
 			}
-			const anyChildIsSelected = termTree.children.map( treeHasSelection ).filter( ( child ) => child ).length > 0;
+			const anyChildIsSelected =
+				termTree.children
+					.map( treeHasSelection )
+					.filter( ( child ) => child ).length > 0;
 			if ( anyChildIsSelected ) {
 				return true;
 			}
@@ -258,7 +305,9 @@ class RadioTermSelector extends Component {
 	setFilterValue( event ) {
 		const { availableTermsTree } = this.state;
 		const filterValue = event.target.value;
-		const filteredTermsTree = availableTermsTree.map( this.getFilterMatcher( filterValue ) ).filter( ( term ) => term );
+		const filteredTermsTree = availableTermsTree
+			.map( this.getFilterMatcher( filterValue ) )
+			.filter( ( term ) => term );
 		const getResultCount = ( terms ) => {
 			let count = 0;
 			for ( let i = 0; i < terms.length; i++ ) {
@@ -269,15 +318,14 @@ class RadioTermSelector extends Component {
 			}
 			return count;
 		};
-		this.setState(
-			{
-				filterValue,
-				filteredTermsTree,
-			}
-		);
+		this.setState( {
+			filterValue,
+			filteredTermsTree,
+		} );
 
 		const resultCount = getResultCount( filteredTermsTree );
 		const resultsFoundMessage = sprintf(
+			/* translators: %d: number of results */
 			_n( '%d result found.', '%d results found.', resultCount ),
 			resultCount
 		);
@@ -297,12 +345,20 @@ class RadioTermSelector extends Component {
 			// Map and filter the children, recursive so we deal with grandchildren
 			// and any deeper levels.
 			if ( term.children.length > 0 ) {
-				term.children = term.children.map( matchTermsForFilter ).filter( ( child ) => child );
+				term.children = term.children
+					.map( matchTermsForFilter )
+					.filter( ( child ) => child );
 			}
 
 			// If the term's name contains the filterValue, or it has children
 			// (i.e. some child matched at some point in the tree) then return it.
-			if ( -1 !== term.name.toLowerCase().indexOf( filterValue.toLowerCase() ) || term.children.length > 0 ) {
+			if (
+				-1 !==
+					term.name
+						.toLowerCase()
+						.indexOf( filterValue.toLowerCase() ) ||
+				term.children.length > 0
+			) {
 				return term;
 			}
 
@@ -319,20 +375,23 @@ class RadioTermSelector extends Component {
 
 		return renderedTerms.map( ( term ) => {
 			const id = `editor-post-taxonomies-${ klass }-term-${ term.id }`; // @helgatheviking
+
+			// ID of the currently selected term. If no terms, select the default.
+			const selected = -1 !== terms.indexOf( term.id ) || ( ! terms.length && term.id === taxonomy.default_term ) ? term.id : 0;
 			return (
-				<div key={ term.id } className="editor-post-taxonomies__hierarchical-terms-choice">
-					<input
-						id={ id } 
-						className="editor-post-taxonomies__hierarchical-terms-input"
-						type="radio" // @helgatheviking
-						checked={ terms.indexOf( term.id ) !== -1 }
-						value={ term.id }
-						onChange={ this.onChange }
-						name={ 'radio_tax_input-' + this.props.slug } // @helgatheviking
+				<div key={ term.id } className={ `radio-taxonomies-choice editor-post-taxonomies__hierarchical-terms-choice ${taxonomy.hierarchical ? "": "editor-post-taxonomies__non-hierarchical-terms-choice"}` }>
+					<RadioControl
+						selected={ selected }
+						options={ [
+							{ label: unescapeString( term.name ), value: term.id },
+						] }
+						onChange={ () => {
+							const termId = parseInt( term.id, 10 );
+							this.onChange( termId );
+						} }
 					/>
-					<label htmlFor={ id }>{ unescapeString( term.name ) }</label>
 					{ !! term.children.length && (
-						<div className="editor-post-taxonomies__hierarchical-terms-subchoices">
+						<div className={ 'editor-post-taxonomies__' + klass + '-terms-subchoices ' }>
 							{ this.renderTerms( term.children ) }
 						</div>
 					) }
@@ -342,19 +401,41 @@ class RadioTermSelector extends Component {
 	}
 
 	render() {
-		const { slug, taxonomy, instanceId, hasCreateAction, hasAssignAction } = this.props;
+		const {
+			slug,
+			taxonomy,
+			terms, // @helegatheviking
+			instanceId,
+			hasCreateAction,
+			hasAssignAction,
+		} = this.props;
+
 		const klass = taxonomy.hierarchical ? 'hierarchical' : 'non-hierarchical'; // @helgatheviking
 
 		if ( ! hasAssignAction ) {
 			return null;
 		}
 
-		const { availableTermsTree, availableTerms, filteredTermsTree, formName, formParent, loading, showForm, filterValue } = this.state;
-		const labelWithFallback = ( labelProperty, fallbackIsCategory, fallbackIsNotCategory ) => get(
-			taxonomy,
-			[ 'labels', labelProperty ],
-			slug === 'category' ? fallbackIsCategory : fallbackIsNotCategory
-		);
+		const { 
+			availableTermsTree, 
+			availableTerms, 
+			filteredTermsTree, 
+			formName, 
+			formParent, 
+			loading, 
+			showForm, 
+			filterValue,
+		} = this.state;
+		const labelWithFallback = (
+			labelProperty,
+			fallbackIsCategory,
+			fallbackIsNotCategory
+		) =>
+			get(
+				taxonomy,
+				[ 'labels', labelProperty ],
+				slug === 'category' ? fallbackIsCategory : fallbackIsNotCategory
+			);
 		const newTermButtonLabel = labelWithFallback(
 			'add_new_item',
 			__( 'Add new category' ),
@@ -372,7 +453,7 @@ class RadioTermSelector extends Component {
 		);
 		const noParentOption = `— ${ parentSelectLabel } —`;
 		const newTermSubmitLabel = newTermButtonLabel;
-		const inputId = `editor-post-taxonomies__${ klass }-terms-input-${ instanceId }`; // @helgatheviking			
+		const inputId = `editor-post-taxonomies__${ klass }-terms-input-${ instanceId }`; // @helgatheviking
 		const filterInputId = `editor-post-taxonomies__${ klass }-terms-filter-${ instanceId }`; // @helgatheviking
 		const filterLabel = get(
 			this.props.taxonomy,
@@ -386,20 +467,33 @@ class RadioTermSelector extends Component {
 		);
 		const showFilter = availableTerms.length >= MIN_TERMS_COUNT_FOR_FILTER;
 
+		const noneSelected = terms.length ? 0 : -1; // @helgatheviking
+		const noneLabel = sprintf( // @helgatheviking
+			/* translators: %s: taxonomy name */	
+			_x( 'No %s', 'term', 'radio-buttons-for-taxonomies' ),
+				get(
+					this.props.taxonomy,
+					[ 'labels', 'singular_name' ],
+					slug === 'category' ? __( 'Category' ) : __( 'Term' )
+				)
+			);
+
 		return [
-			showFilter && <label
-				key="filter-label"
-				htmlFor={ filterInputId }>
-				{ filterLabel }
-			</label>,
-			showFilter && <input
-				type="search"
-				id={ filterInputId }
-				value={ filterValue }
-				onChange={ this.setFilterValue }
-				className="editor-post-taxonomies__hierarchical-terms-filter"
-				key="term-filter-input"
-			/>,
+			showFilter && (
+				<label key="filter-label" htmlFor={ filterInputId }>
+					{ filterLabel }
+				</label>
+			),
+			showFilter && (
+				<input
+					type="search"
+					id={ filterInputId }
+					value={ filterValue }
+					onChange={ this.setFilterValue }
+					className="editor-post-taxonomies__hierarchical-terms-filter"
+					key="term-filter-input"
+				/>
+			),
 			<div
 				className="editor-post-taxonomies__hierarchical-terms-list"
 				key="term-list"
@@ -408,6 +502,21 @@ class RadioTermSelector extends Component {
 				aria-label={ groupLabel }
 			>
 				{ this.renderTerms( '' !== filterValue ? filteredTermsTree : availableTermsTree ) }
+
+				{ taxonomy.radio_no_term && (
+					<div key="no-term" className={ 'editor-post-taxonomies__' + klass + '-terms-choice ' }>
+						<RadioControl
+							selected={ noneSelected }
+							options={ [
+								{ label: noneLabel, value: -1 },
+							] }
+							onChange={ () => {
+								this.onClear();
+							} }
+						/>
+					</div>
+				) }
+
 			</div>,
 			! loading && hasCreateAction && (
 				<Button
@@ -446,7 +555,7 @@ class RadioTermSelector extends Component {
 						/>
 					}
 					<Button
-						isDefault
+						isSecondary
 						type="submit"
 						className="editor-post-taxonomies__hierarchical-terms-submit"
 					>
@@ -455,7 +564,6 @@ class RadioTermSelector extends Component {
 				</form>
 			),
 		];
-		/* eslint-enable jsx-a11y/no-onchange */
 	}
 }
 
@@ -465,9 +573,25 @@ export default compose( [
 		const { getTaxonomy } = select( 'core' );
 		const taxonomy = getTaxonomy( slug );
 		return {
-			hasCreateAction: taxonomy ? get( getCurrentPost(), [ '_links', 'wp:action-create-' + taxonomy.rest_base ], false ) : false,
-			hasAssignAction: taxonomy ? get( getCurrentPost(), [ '_links', 'wp:action-assign-' + taxonomy.rest_base ], false ) : false,
-			terms: taxonomy ? select( 'core/editor' ).getEditedPostAttribute( taxonomy.rest_base ) : [],
+			hasCreateAction: taxonomy
+				? get(
+						getCurrentPost(),
+						[ '_links', 'wp:action-create-' + taxonomy.rest_base ],
+						false
+				  )
+				: false,
+			hasAssignAction: taxonomy
+				? get(
+						getCurrentPost(),
+						[ '_links', 'wp:action-assign-' + taxonomy.rest_base ],
+						false
+				  )
+				: false,
+			terms: taxonomy
+				? select( 'core/editor' ).getEditedPostAttribute(
+						taxonomy.rest_base
+				  )
+				: [],
 			taxonomy,
 		};
 	} ),
@@ -478,5 +602,4 @@ export default compose( [
 	} ) ),
 	withSpokenMessages,
 	withInstanceId,
-	//withFilters( 'editor.PostTaxonomyType' ), // intentionally commented out
 ] )( RadioTermSelector );