+++ public/wp-content/plugins/disable-emails/changelog.md
@@ -2,6 +2,43 @@
 
 ## Changelog
 
+### 1.8.2, 2023-11-22
+
+* fixed: deprecation warnings in PHP 8.1+
+
+### 1.8.1, 2022-05-26
+
+* fixed: fatal exception when an email address was invalid
+
+### 1.8.0, 2022-02-16
+
+* fixed: mock PHPMailer object did not have recipient addresses for developers to inspect (thanks, [SpartakusMd](https://wordpress.org/support/users/spartakusmd/)!)
+
+### 1.7.0, 2020-08-11
+
+* fixed: WordPress 5.5 compatibility
+
+### 1.6.3, 2020-03-19
+
+* fixed: unhandled `phpmailerException` exceptions (thanks, [y0uri](https://wordpress.org/support/users/y0uri/)!)
+
+### 1.6.2, 2020-03-10
+
+* fixed: activating the must-use plugin throws an error if the mu-plugins folder is missing
+* changed: can now enable both a Toolbar indicator and a site-wide admin notice when emails are disabled
+* changed: filter hook `disable_emails_indicator` also accepts 'notice_toolbar' to enable both notice and Toolbar indicator
+
+### 1.6.1, 2019-12-21
+
+* fixed: Toolbar indicator has no visible slash in Sunrise admin theme
+* fixed: undefined function `is_plugin_active_for_network()` (thanks [isabelc](https://github.com/isabelc)!)
+
+### 1.6.0, 2019-12-15
+
+* fixed: undefined index for `$_SERVER['SERVER_NAME']` when emails sent during wp-cli
+* added: indicator setting to show either a Toolbar indicator or a site-wide admin notice when emails are disabled
+* added: filter hook `disable_emails_indicator` for setting the indicator from code; accepts 'none', 'toolbar', 'notice'
+
 ### 1.5.0, 2019-11-11
 
 * fixed: PHP notice -- Trying to get property 'ErrorInfo' of non-object